requirements.txt
google-genai>=1.0.0
chromadb>=1.0.0
gradio>=6.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
socksio>=1.0.0
plotly>=5.0.0
datasets>=2.0.0
scikit-learn>=1.0.0
